Cope with a Type 1 lazily derived from another Type 1
In this case, the definition of the "narrow bold" variant contains the entire
data stream defining the "bold" font from which it then picks the CharStrings.

This case only work fully as intended when the font is interpreted by a full
PostScript interpreter.

In this case, the "narrow" aspect comes from the FontMatrix scaling the x
axis down very slightly. Other than that, the various names and the UniqueID
the two fonts are identical.

To cope with this, we now only allow the FontMatrix, name fields and UniqueID
(and UID) to be set once.