PS interpreter - improve setcolorspace and friends
This arose (I think) from bug #707990, where the colour space array
argument to setcolorspace is modifed by the tint transform procedure
during the course of setcolorspace or currentcolor.

The basic idea here is to make a copy of the colour space array after
we have validated the colour space array but before we begin processing
it. After that the C code will use the copy of the array, not the
original colour space array as defined in PostScript, for all operations
which prevents this kind of abuse.

We don't need a totally general purpose function to do this, because
colour space arrays must conform to limited criteria, and we don't
copy the array until after it has been validated. We do need to take
care to preserve the VM allocation mode and attributes of any compound
objects.

We deliberatly do not copy the tint transform procedure itself, it is
permissible (mad, but legal) for that procedure to modify itself, so
we continue to allow that.

We would like to limit the use of the tint tramsform to setcolorspace,
which at first sight should be possible, because we convert all the
tint transforms to either a PostScript calculator function or, for
complex cases, a sampled function at the time we set the space. The
graphics library uses those functions, not the PostScript procedures,
when needed.

Unfortunately, as noted in septransform and devicentransform, Adobe
Photoshop defines DeviceN spaces with an illegal tint transform (it is
illegal because it has side effects). We can convert that to a function
but if we run it we get an error. We must run the tint transform (when
not producing the inks), so this is a complete blocker and means that
we must run the PostScript tint transform for setcolor.

We don't however, need to run it for the current(gray|rgb|hsb|cmyk)color
operators, and this commit removes the framework for that, replacing it
with much simpler and more robust code.

The hacky code in currentcolorspace, for working around yet another kind
of broken PostScript by Adobe has been replaced with simpler code that
doesn't require us to execute PostScript.