pdfimage24 devices - Fix JPEG height when image type is JPEG
Bug #708255 - "pdfimage24 -dDownScaleFactor creates distorted aspect ratio (too tall) with -sCompression=JPEG"

This only affects JPEG because that's the only supported file *format*
that can be embedded in a PDF file (we don't produce JPX output). The
PDF is correct, but the JPEG height marker is incorrect, it uses the
device height (without accounting for downscaling).

Looks like a simple oversight, fixed here.