ps2write - emit some controls before ProcSets when ProduceDC is false
Bug #708257 " ps2write with -dProduceDSC=false produces broken file"

Not sure why this is different to the DSC version, doesn't seem to be
any reason. Its the whole content though, not just the 'end', so move
all the controls as well as the end before the emission of the ProcSets.