HPGL/2 - better handle SI with proportional fonts and LO
Bug 708135 "Label origin 7 in HP/GL doesn't work for proportional fonts"

In fact label origin 7 works correctly here, what doesn't work is the
'SI' (absolute character size) when combined with a propertional font.

I *think* (and it isn't totally clear) that when SI is in operation all
characters are intended to be forced to a specific width and height.

For a fixed pitch font this isn't too onerous, we just alter the CTM
until the width and height match the target.

For a proportional font we would need to do this individually for every
glyph, and the code in hpgl_print_char makes no attempt to do so; it
simply assumes a particular width and scales the CTM to make that fit.
So all the characters are not the same width....

Ideally we should fix hpgl_print_char, but I can't see any way to do
that (probably due to unfamiliarity with the code). However it seems
that this isn't a huge problem anyway, given that nobody has ever
reported it before.

But what is being perceived as the bug isn't the fact that all the
characters are differeing sizes; it's the fact that the 'label origin'
is being calculated assuming that all the characters are the same size.
Since they are not, the origin is calculated incorrectly and leads to
text being misplaced.

So in this commit, add some hackery.... We figure out the width of the
string using the actual character widths. Then we figure out what
scale factor we would use for the absolute positioning, and the scale
factor is we were just drawing normally.

What we want is the width that will actually get drawn, we know the
widths now, and we know the scale factro that was used when those widths
were retrieved. If we divide by the 'not set' scale factor we get the
'unscaled' width, and then multiplying by the scale which will be
applied when drawing gives us the actual width.

We then shift the origin by that amount for the various label origin
operations.

The net result is that we end up drawing the text in the right place to
fill the area and end at the correct location.

So not a fix, but a reasonable work around for what I think is a very
niche use.