pdfwrite - provide means to set Flate parameters for image compression
Bug #708123 "Allow setting zlib flate compression level to max"

The Adobe Acrobat Distiller parameters allow the user to configure
the DCT (JPEG) settings for imaeg compression by using
(gray|Color)ImageDict or (Gray|Color)ACSImageDict, and for JPX encoding
(which we don't support) using JPEG2000(Gray|Color)ImageDict and
JPEG2000(Gray|Color)ACSImageDict, but there is no way to set the Flate
compression parameters.

We could opt to have another dictionary, as Adobe have done for JPX,
but that's verbose. Instead, we note that the Flate and JPEG Keys are
unique, there is no duplication. This allows us to extend the
(Gray|Color)ImageDict and (Gray|Color)ACSImageDict to include the two
parameters of use to Flate; /Predictor and /Effort.

This commit does exactly that, it is now possible to specify parameters
for both JPEG and Flate compression and have them apply either to all
images (ImageDict) or to the compression selected by the auto-chooser
(ACSImageDict). When using the auto-chooser only the keys relevant to
the chosen compression are used.