pdfwrite - Produce PDF/X-1a output
Bug #691585 "Need PDF X/1-a format output"

This commit alters the PDFX switch from a boolean to an integer, so that
we can have different flavours of PDF/X. This means changes throughout
the pdfwrite device, and also to gs_pdfwr.ps to cater for the change.

After that we move on to adding support for PDF/X-1a. Broadly this is
the same as PDF/X-3 but the Page object must now have a /ArtBox and we
can't support DeviceN colour spaces.

Changed a number of messages from PDF/X-3 to just PDF/X.

A number of places were setting PDF/X to 0 when they *should* have been
setting PDF/A to 0, because we are aborting PDF/A production, not PDF/X
production. Fixed those.

We were accidentally permitting PDF/X files to omit the ID, that is not
allowed, and has been fixed.

Updated the documentation to ntoe that we can produce PDF/X-1

Updated the example pdfx_def.ps file to emit the correct identifier
when emitting PDF/X.

There is allegedly an updated version of the specification which uses
PDF 1.4 instead of PDF 1.3 (but does not allow any new features) but
Acrobat won't validate my attempts at that. As a result I haven't
added support for that variant.