Bug 708277: Avoid ioerror on empty "PostScript" file
Commit:

58d93d87dfc0edaffdde8b093df8acc11a6a6066

addressed an UBSAN complaint about setting a pointer value to "NULL - 1", when
the file has zero length, and thus doesn't create a memory buffer in the
stream object. But it resulted in a ioerror when trying to read from such an
empty file.

This change avoids the problem by using a dummy, zero length buffer to cope with
that situation.