pdfwrite - implement PDF/X-4 output
Bug #704830 "Feature request: Implement PDF/X-4 output"

As the feature request says; implement PDF/X-4 output.

This is, obviously, a new feature and some teething troubles are to be
expected, especially in the area of DeviceN colour spaces.

Also, as part of the commit it was necessary to alter the XMP metadata
which has meant adding a coulpe more entries. A few test files which
produced exceedingly small output files have grown in size by > 5%
simply because the metadata is such a large portion of the output.