Fix the stream dummy buffer workaround.
UBSAN complained because the stream code initialises the cursor buffer pointer
to one byte behind the actualy buffer address. But there are cases where that
buffer pointer is NULL, and assigning "NULL - 1" is undefined.

The original fix:
58d93d87dfc0edaffdde8b093df8acc11a6a6066

avoided the problem by special casing the NULL pointer case, and simply
using NULL (rather than NULL - 1). But this caused an undesireable ioerror when
reading an empty file.

To work around that, the second fix:
8d7f8b9702d908b0f113e22d701ca2d7fbab71a3

was to use static, dummy, empty buffers
one for read and one for write cursors. This causes a problem because of how
the stream cursor union overlaps values. Read cursors end up with the buffer
start pointing the read dummy buffer, and the buffer limit pointing to the
end of the write dummy buffer.

So, make the two cases (read and write) use the same dummy buffer.

The dummy buffer is one byte (rather than zero bytes) because some versions of
Visual Studio won't accept a zero length array.

oss-fuzz 393414239, 393411521, 393408793