pdfwrite - alter the emission of multi code point ToUnicode CMaps
Bug #708284 "Cant find text and selected text is pasted as symbols"

The actual fault in the bug report is simply that the ToUnicode CMap in
the PDF file is incorrect.

Previous work surrounding ToUnicode CMaps with multiple code points had
led us to drop the entire CMap when we encountered a single such entry.

We still can't deal with ToUnicode CMaps on input which have more than 2
UTF-16 code points, and until we can this problem cannot be addressed
'correctly'. However, if we only drop the ToUnicode entry, rather than
the entire CMap, then we can do a 'better' job for files of this kind,
even though they are technically invalid.

So here we do away with the 'TwoByteToUnicode' flag in the pdfwrite
font structure. Instead we rejig pdf_add_ToUnicode so that we drop
entries which might not work, instead of dropping the entire map.