Bug 708282: Generate pseudo XUIDs for copied pdfi fonts
Originally, we made copied font UIDs invalid, since reusing the original could
cause cache problems if the encoding, widths or certain other aspects of the
font were changed in the PDF font object.

But then this causes pdfwrite to have difficulty differentiating different fonts,
with certain poorly created PDF inputs.

Since the generation of pseudo XUIDs uses the object number of the font, we
can generate a _new_ pseudo XUID for the copied font, allowing pdfwrite to
differentiate the fonts, and avoid the original cache problems.