Bug 708101: pdfi: Generate fake XUID even without a filename
Normally, we use a hash of the input file name as part of our pseudo XUID
creation, but if the intput is read from stdin (so interpreted from a temp file)
there is no file name.

So, in that case, use some of the contents of the stream buffer to generate the
hash.

In the case of the bug file, the lack of an input file name meant we didn't
created the fake XUID, which removed one of the ways that pdfwrite identifies
fonts, meaning it mis-identified a font, and caused the "corrupt" output.