pdfwrite - fix a typo leading to potential uninitialised memory use
Arising from bug #708101; Chris noticed that memory sanitiser was
complaining about uninitialised memory use. The reasons for this are
complex and actually fixed by his own commit to resolve the reported
problem, but it exposed an underlying problem.

We were using the wrong variable (pdfsubf instead of pdfsubf0) and
pdfsubf could, potentially, be uninitialised in certain error
conditions (as exposed by this bug).

This commit just changes to testing the correct variable. It lloks like
this had been incorrect for quite some time.