pdfwrite - fix a ubsan warning
The arithmetic we were using caused conniptions for ubsan. This code
simply uses the top bit of the gs_offset_t as a flag; if it is set then
the position is in the 'asides' file, if it isn't then it is a position
in the main file.

Obviously we reset the bit before trying to use the offset.

Getting this cast in a form ubsan and the compiler were both happy
with, and capable of dealing with 32-bit offsets, took a little bit of
trial and error.