pdfwrite - don't use inline colour space names for base of complex spaces
This turned up while testing release candidates. With the file

/tests_private/pdf/sumatra/multiply_luminosity_masks_ignored.pdf

There is a small image in a DeviceN space, which is actually an
NChannel space. Because the image is small we try to write it inline,
but we fail because only simple colour spaces can be used for inline
images.

However, we set up the 'base' space names to use the inline
abbreviations, so when we come to writing out the base space of the
NChannel space we were accidentally writing /CMYK instead of /DeviceCMYK
which caused Ghostscript to emit a warning on the output file.

Add some more checking to make sure we use the correct names.