PDF interpreter - don't create fake XUID for disk based fonts
Bug #708291 "dramatic slowdown and size increase of output files"

Lilypond is sending PDF files with multiple fonts, which are not
embedded in the input, and using Ghostscript to embed the fonts in the
output.

Additionally it is relying on the substitute fonts being recognised as
the *same* font, so that we don't embed multiple fonts.

We've had to work hard over the years to make sure that fonts which are
*not* the same are not identified as being the same, as this causes
incorrect output. Commit 	3276619a7eb5b2269a158f115a163847be31458e
added fake XUIDs for 'copied' fonts but this breaks Lilypond's usage.

This commit only adds the fake XUID to copied fonts if they are *not*
disk-based fonts (the PDF font object has a NULL filename). We assume
that fonts loaded from disk are always the same font if they have the
same name.