PDF interpreter - catch direct Info dictionaries and warn
Uncovered by release testing. The Info dictionary is specified as being
required to be an indirect object in the trailer dictionary. We have
(of course!) several examples of PDF files where the Info dictionary
is instead directly defined in the trailer.

This was causing the loop detection to go awry, as we were adding object
number 0 to the loop detector.

Catch the underlying problem here and issue a warning.