PDF interpreter - prevent adding object number 0 to the loop detection
This follows on from the invalid Info object in the trailer dictionary.
That was adding object 0 to the loop detector, which we must not do as
we use object 0 (the head of the linked list of free objects) as a
'mark' in the loop detector.

Rejig the code so that we don't add object 0 and warn if code tries to.