devices - use correct index when sorting fonts
This came to light frm work Chris was doing which will be committed
later.

If cfdata->num_glyphs does not match the actual number of copied glyphs
with the 'used' flag set then we could attempt to qsort more names
than we actually copied, leading to a potential crash.

By using the actual number of names copied we avoid the crash. The
value of cfdata->num_glyphs *should* be the same as the actual number of
'used' glyphs, but that's a separate problem and will be addressed in a
following commit.

Also, noticed while working on this, an error condition could lead to a
memory leak; fix that too.