devices - update glyph number correctly in copied fonts
This follows on from the previous commit and is again exposed by work
Chris is doing with fonts.

If we 'uncopy' a glyph from a copied font we need to count down the
number of copied glyphs in the font. Failing to do so is what caused
the mismatch between 'num_glyphs' and the number of glyphs with 'used'
set. That caused potential seg faults which were fixed by the prior
commit.

In addition; when we 'uncopy' a glyph we should set the 'data' pointer
to NULL. There are various places which use this as an indicator that a
glyph is valid instead of using either the 'size' or 'used' members.

Failing to set this to NULL can lead to memory corruption.