PDF interpreter - don't error out when checking the first XObject fails
Bug #708302 "Incorrect results when converting PDF to image"

The problem occurs because the file has pages which contain Resource
dictionaries which reference XObjects which are defined as free objects.

Mostly this doesn't matter, but if the *first* XObject we check is a
free object then we fail to check the remaining ones, which leads to a
failure to detect transparency.

Move the object type checking into the main loop so that we don't exit
if the first object is not the correct type.

Do the same for Shadings and Patterns; Fonts and Colour Spaces appear
to permit this already. We don't give a warning about incorrect types
(though we do for accessing a free object), if the objects actually
get used then we'll get a warning or error when processing the page.