PDF interpreter - fix buffer overrun in CFF font XUID scanning
OSS-fuzz issues 399174690, 399173691, 399183998, 399390078, 399388654,
399390672 at the time of writing this message

The loop writing xvalues could overrun the buffer; when i is 0 and n
is 8 we write to xvalues[8], but valid values are 0-7.

Just correct the indices.

This may not fix all the reported issues but they all look the same to
me, and it certainly addresses many of them.