PDF interpreter - fix CFF font XUID parsing with an empty XUID
OSS-fuzz #399388654

The fuzzed file has a font with an XUID, but the XUID has no bytes. This
caused us to allocate a byte array with nothing in it and, for reasons
I haven't tracked down, causes memory corruption.

On OSS-fuzz this resulted in a heap-us-after-free but for me on both
Windows and Linux it causes a signal 11 crash.

This commit simply sets the XUID to invalid id the length is 0.