PDF interpreter - CFF fonts and UIDs/XUIDs
OSS-fuzz #400002189

The problem here is that we have a CFF font which has a UID, but the
value of the UID is invalid, less than 0, which is a marker for an XUID.

Because it 'looks like' an XUID, when we come to copy the uids we try
to copy the xvalues associated with the XUID, but since we never
assigned xvalues (because it isn't an XUID) we end up trying to
reference location 0.

This commit simply ignores a UID value in the XUID range, leaving the
UID set to 'invalid'. While passing, fix a couple of compiler warnings
about casting double to float.