PDF interpreter - make stderr/stdout usage more consistent
As a result of looking into a customer problem I noticed that, despite
the comment in base/std.h:

 * Use dmprintf (and dmputs, dmputc, dmflush etc for messages intended for
 * stdout). If you would like an optional filename/line number (where

dmprintf in fact goes to stderr. Not only that, but it does not respect
-q or -dQUIET.

We were using dmprintf quite extensively in the PDF interpreter which
meant that a lot of output went to stderr when it should have gone to
stdout, and messages which should have been suppressed with -q were
in fact still being printed.

I've reviewed all the uses of dmprintf in the code and replaced all of
them.

For information dumping (eg -dPDFINFO) we now use outprintf so that
the output goes to stdout.

For debug prints (-dPDFDEBUG) we now use outprintf so that the output
goes to stdout.

For debug-time warnings to developers (such as unimplemented features) we
now use dbgprintf.

A few places have gained our normal error/warning strategy where they
printed to stderr before.

I'm sure more improvements could be made but hopefully we are at least
more consistent throughout now.