pdfwrite - set AutoFilter to default to true
This has no effect on Ghostscript, because the PostScript startup
framework already defaulted AutoFilter to true, but for the other
interpreters (gpdf, gxps, gpcl and gpdl) this means that the image
compression chooser will be activated.

This does result in some differences in the output, because we may now
use DCT encoding instead of Flate encoding, or vice-versa. However since
it is tricky to set this from the command line I think its preferable to
have it default the same as Ghostscript.