pdfwite - fix implementation of Flate ACSImageDict
I missed this when implementing Flate parameters in (Color|Gray)
ACSImageDict parameters. The code tries to set the DCT filter using the
parameters and only sets up the dictionary if that doesn't produce an
error.

That isn't going to work going forwards as we could have DCT or Flate
now, and potentially Brotli parameters in future.

So just remove the (probably pointless) checking.