pdfwrite - fixes for auto compression chooser
This was noticed while testing the new Brotli encoding. Two fixes:

1) The compression chooser counts 'plateaus' (I think runs of identical
pixels) to try and decide if an image is photographic or linework. The
problem here was if both 'plateaus' and gradients were 0 then we
incorrectly identified an image as linework, leading to us using the
lossless compression instead of DCT.

2) While writing image data, if one of the streams (lossless or lossy)
is very much smaller than the other then we abort and use that scheme.
But we were not accuonting for one of the streams hitting that margin
while the other stream had not yet written any bytes.