PDF interpreter - add a warning for broken image Matte
Robin pointed this out while working on a customer problem; if the Matte
handling returned an error we silently ignored it and proceeded,
ignoring any Matte entry.

We should really note a warning, and stop if directed to do so, this
commit just adds that to the existing handling. If not directed to stop
we will dimply continue as before, ignoring any Matte entry.