PDF interpreter - fix Coverity warning
Coverity ID #457490

It was possible for fname to be used by memcpy when it was NULL if
fname is initially NULL and pdfi_open_resource_file_inner() doesn't
allocate 's'.

We've been fixing these because of ubsan warnings, I guess this is one
we hadn't found yet but Coverity has noticed the code changes and
started picking up on them.