Reinstate pdf2dsc
After feedback from users (AUCTeX and gv) put back the pdf2dsc utility
but note in the comments that this is now unsupported code (in truth
I think it always was, but this makes it explicit).

Because the PostScript program uses undocumented parts of the old
'written in PostScript' PDF interpreter portions of it probably don't
work and it may fail altogether at some point.