Bug 708319: Further fixes to tag devices.
Tweak the pdf14 device setup to be better (or at least clearer)
in the way it handles tags.

The actual important fix for this file is the correction in the
calculation of num_planar_planes; repeated pops of color models
in the presence of tags were causing this number to fall to zero,
at which point we were using chunky rendering rather than planar.