Improve clipper device reference counting
OSS-fuzz #401846983

The problem here is complicated, but eventually it comes down to the
fact that the 'clipper' device is a forwarding device and, as such, has
a reference to a target device, but it does not increment the reference
count of that object when it takes a reference to it.

We have a Pattern, which fills an area using another Pattern, and that
pattern declares itself so large it uses a pattern clist. In addition
the child pattern uses a Form XObject which has a transparency Group.

When we fill with the first pattern we start by creating a clipper
device, created 'on the stack', ie a local device structure, not an
allocated piece of memory. That device gets set up with a 'target'
pointer which points to the pattern accumulator. We then run the pattern
PaintProc to fill the bitmap tile.

That causes us to run the second pattern, which gets set up as a pattern
clist accumulator. This is important because the pattern tile cache has
a reference to the device, which does not exist if we run a normal
pattern.

We again run the paintproc for the subsidiary pattern, which causes us
to push a compositor. When we pus the compositor it inserts itself
between the 'clipper' and 'pattern accumulator' devices. It properly
reference counts the clipper->target but, because the clipper device
didnt' count *up* the target, this overall results in the count going
down by one.

When we flush the pattern tile cache that ends up decrementing and
freeing the pattern accumulator, because its reference count is one too
low, even though there is still a reference to it (from the enclosing
transparncy Group), which eventually causes the crash.

We get away with this most of the time, but in this particular case
circumstances conspire to expose the problem.

Ideally we should have the 'clipper' device properly reference count
its 'target', but there are a lot of places that use it, and some are
complicated to trace teh lifetime.

As an interim measure this commit uses the existing 'destroy' function
to count down the target, and explicitly counts it up in a couple of
places where we use the 'destroy' function. It also fixes the one
existing location where we were using the destroy funcion; there were a
number of ways to exit the function without calling 'destroy', which
was acceptable when all it did was meddle with the cached clip, but
is insufficient when we are going to play with reference counts.