Bug 708235(redux): Don't use resource search paths for native fontmap
A misunderstanding of what the old PostScript code used to do led me to use the
resource search paths (default, and given by the "-I" command param) to search
for font files to add to the native fontmap in pdfi.

In fact, the PS code does not do that. It searches the -I paths for PS style font
_resource_ (resources are essentially identified by file name) when a
non-embedded is requested.

It then checks the fontmap, and after that the native fontmap.

The native fontmap being assembled by scanning the paths in FONTPATH for files
that we can identify as fonts.