Fix clipper device reference counting
This follows on from commit 7588c1d8c5ba3379dc4f739eb98a6381661441ef

That commit fixed a problem with the clipper device failing to
reference count its 'target' member, but only fixed the specific
instance uncovered by OSS-fuzz.

It's clear that there are other places that will have the same problem
(fill+stroke for example) and many other places that *migh* suffer from
the same issue; the clipper device's target being replaced.

This commit counts up the 'target' reference count when clipper devices
are created on the stack or heap, counts down the target in the
'destroy' function and in the finalize function, and makes certain that
every call to create a clipper device is matched with a call to the
'destroy' function (or is finalized).

This should prevent any more occurences of the same problem with this
device.