Suppress Coverity 457989
ThIs is actually a Coverity false-positive, the tests for imaging_dev
being equal to &cdev mean that we can't execute the detected code that
uses cdev->target without first calling gx_make_clip_device_on_stack,
which initialises cdev.target and cdev.cpath.

But just to silence it, initialise a couple of members of a structure.