Bug 708319 continued: Attempt to improve speed.
A significant amount of time appears to be spent rendering paths
through clips.

The way we do this is to keep a list of rectangles that makes up
the clip. We then set up a clipping path that represents just a
single one of these rectangles in turn, and plot the path through
it.

The setup of this single rectangle clipping path can be faster, so
we add some dedicated functions for this.

Also, we were spending 10% of the time fetching ids, largely due
to the locking. Fetch 100 at a time.