Bug 708336: Initialise the jmp_buf before calling setjmp()
This is based on repeated investigations and considerable guess work.

In rare circumstances, a gs binary built with clang (various versions 12-18,
so far) will crash on error return from the libjpeg code. For most error returns
from the libjpeg code, we rely on setjmp/longjmp.

The precise location and nature of the crash varies, and is due to the stack
frames being corrupted on the error return from setjmp.

As part of the investigation, I found that setting the entire jmp_buf to zeros
before the setjmp call(s) caused the problem to no longer occur. Closer
inspection shows the stack frames are not being corrupted in the same manner.

Since I've not (yet) been able to reproduce the problem in a simple context
suitable for reporting upstream, and initialising the buffer appears to work,
I've opted to push this change.

This is definitely a workaround, rather than a fix.