HP/GL-2 interpreter - PU don't append a point to path if pen is already up
Bug 708299 "Squares rendered when sending a HP/GL 'PU'"

The bug is much more complex than the description in the report, and
while it is true that the PU command is appending a point, it is also
necessary for there to be an existing path, with no points in it. In the
example file this happens because of the CI (circle) command which has
an implicit PD (pen down) command.

The CI command draws a circle, regardless of the state of the pen, the
documentation says it has an implicit pen up and returns the pen to the
prior state at the end of the circle. However, that means we have to
reset the path. Doing that leaves us with an empty path in a particular
state.

The PU command is implemented by using the hpgl_plot() command, which
will add a point to the path. If the path is in the state after a circle
then this results in the plot code drawing that point, because it adds
the point before drawing, so it decides it needs to draw the pooint.

We can avoid that by pretending to be a 'PA' because the PA command,
with no parameters, does not append a point.

But that causes a problem if we do a PU with parameters, because in that
case we *should* append a point, the point we moved to before the pen
was lifted.

Of course this only matters if the pen was not already up, if it was
then we don't draw a path anyway. So we can use that to determine
whether we want to append a point or not. If the pen was down, then
we do append a point, if it was up, then we don't append a point.

This causes a small difference in one test file, I suspect that it was
wrong before and nobody noticed because the only effect is to make
some dashes in a line slightly larger.