PDF interpreter & pdfwrite - fix Optional Content for Form XObjects
Bug #708420 "Possible Regression - optional content (layer) creates corrupt output on pdfwrite device"

The problem is caused by the (relatively) new feature to preserve
Optional Content from the input, in the output file.

Forms are handled specially in pdfwrite, and any Optional Content
relating to a form needs to be passed on via a spec_op. This was all
coded, but the pdfwrite code was using an out of date method to generate
the name for a named object (the named object is what we reference from
the Form XObject).

We were using the old, simple 'ObjxxG0' format for the name, but we
were forced to extend that due to bug 708127, if we have multiple input
fiels we cannot simply use the object number, because the number might
be used for different purposes in each file.

We now use a hash dereived from the current input file's name as part
of the name for a named object, which improves the uniqueness. However,
the object name we want is generated in both the PDF interpreter and
the pdfwrite device, and the pdfwrite device can't know the input file
name.

So this commit reworks the spec_op, we now create the name (lavel) for
the named object and pass that to the pdfwrite device, instead of simply
passing the object number. This means the pdfwrite device needs to take
a copy of the name, and manage the lifetime of the assigned memory which
complicates the commit somewhat.

As noted previously, this is a comparatively new feature an may yet
need more work.