pdfwrite - don't warn on /Dest beyond current page when there are no pages
Bug #708444 "-sDEVICE=pdfwrite causes error when splitting PDF with Ghostscript [10.04]"

The PDF file has an OpenAction which we now preserve, the OpenAction
refrences a page (page 1), but when writing one-file-per-page output
we reach the end of file with an unused page. This causes a (not
unreasonable) error that the destination page is larger than the number
of pages in the file.

This commit simply ignores the problem if 'next page' is 0 or the
output type is file_per_page.