pdfwrite - more work on embedding files
Bug #708469 "Embedding a document yields a /rangecheck error"

This is because PDF/A-3 requires us to have a /AF in the Catalog, which
meant we had to generate one if there wasn't one provided in the input.

But of course, we only need to do this for PDF/A-3, non PDF/A output
need not have a /AF in the Catalog as it is optional.

While testing this I noticed that the fix for bug #708066 broke the
files in bug #707725 (this is a complex area and difficult to test) so
the commit here fixes that by restoring some code.

The code now works (as far as I can tell) with the files from all 3
reports.