PDF interpreter - fix logic checking for Key existence
Bug #708487 "NULL pointer dereference in ghostpdf.c"

The problem is simply that the logic to check for the existence of a
key, and the type of its value, didn't check for errors, causing us to
follow the incorrect (object present and of the correct type) path when
an the type was wrong.

Checked all other uses of pdfi_dict_knownget_type() which turned up a
second case in ghostpdf.c but no others.