graphics library - prevent arithmetic overflow with mask clipper device
Bug #708483 "NULL pointer dereference in the in gsroprun1.h"

The point where the error exhibits is nowhere near the point where it
actually occurs, sadly.

The problem here is that we get an overflow of a data type setting up
the mask clipper device for a type 3 image. This causes us to create an
intermediate 'image1' device which is smaller in width than the source
image.

The interaction between the three devices (original image, mask and
final image) is complicated, but the problem occurs because we use the
original width for the mask device when copying data from the source
to the destination. But the array of pointers pointing to the scan lines
is stored *after* the scan line data. The mismatch in widths between
the source/mask devices, and the final device, means that we end up
writing to what the final device thinks is the array of pointers, when
the source and mask devices think its just data.

This later leads to us trying to dereference the data as a pointer
which fails badly.