PDF interpreter - suppres spurious debug warning
Noticed while debugging, if we have an XRefStm then we manufacture a
dictionary for it, which has an object number of 0 (it is 'inline')
which causes the loop detector to throw a warning.

Add some extra code to avoid trying to add an object number of 0,
because that is the 'mark' for the loop detector stack, and we never
add object 0 (except, obviously, when making a mark).